+++ node/lib/Worker.d.ts
@@ -128,6 +128,10 @@ export declare class Worker extends EnhancedEventEmitter {
      */
     get closed(): boolean;
     /**
+     * Whether the Worker died.
+     */
+    get died(): boolean;
+    /**
      * App custom data.
      */
     get appData(): any;
@@ -167,6 +171,6 @@ export declare class Worker extends EnhancedEventEmitter {
      * Create a Router.
      */
     createRouter({ mediaCodecs, appData }?: RouterOptions): Promise<Router>;
-    private died;
+    private workerDied;
 }
 //# sourceMappingURL=Worker.d.ts.map